Move make_tso_specmodel to datamodels.utils module
