+++ jwst/extract_1d/tests/conftest.py
@@ -5,8 +5,8 @@ from astropy.io import fits
 from astropy.table import Table
 
 from jwst.assign_wcs.util import wcs_bbox_from_shape
+from jwst.datamodels.utils.tso_multispec import make_tso_specmodel
 from jwst.exp_to_source import multislit_to_container
-from jwst.extract_1d import extract as ex
 
 
 @pytest.fixture()
@@ -644,7 +644,7 @@ def make_tso_spec_model(n_spectra=10):
         spec_model = make_spec_model(name=f"slit{i + 1}", value=i + 1)
         spec_list.append(spec_model)
 
-    tso_spec = ex.make_tso_specmodel(spec_list)
+    tso_spec = make_tso_specmodel(spec_list)
     model = dm.TSOMultiSpecModel()
     model.spec.append(tso_spec)
     return model
